# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flower delivery shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  category: string;
  inStock: boolean;
  isFeatured: boolean;
  rating: number;
  reviewCount: number;
}

export interface ProductInput {
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  category: string;
  inStock?: boolean;
  isFeatured?: boolean;
}

export interface CatalogSummary {
  totalProducts: number;
  categories: string[];
  minPrice: number;
  maxPrice: number;
  featuredCount: number;
}

export interface OrderItem {
  productId: number;
  productName: string;
  quantity: number;
  price: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  delivering: 'delivering',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  deliveryAddress: string;
  deliveryDate: string;
  deliveryTime: string;
  /** @nullable */
  notes?: string | null;
  status: OrderStatus;
  totalPrice: number;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
}

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  deliveryAddress: string;
  deliveryDate: string;
  deliveryTime: string;
  notes?: string;
  items: OrderItemInput[];
}

export interface Review {
  id: number;
  /** @nullable */
  productId?: number | null;
  customerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  /** @nullable */
  productId?: number | null;
  customerName: string;
  rating: number;
  comment: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiRecommendRequest {
  message: string;
  history?: ChatMessage[];
}

export interface AiReply {
  reply: string;
  recommendedProductIds: number[];
}

export type ListProductsParams = {
category?: string;
minPrice?: number;
maxPrice?: number;
search?: string;
sortBy?: ListProductsSortBy;
};

export type ListProductsSortBy = typeof ListProductsSortBy[keyof typeof ListProductsSortBy];


export const ListProductsSortBy = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  name: 'name',
  popular: 'popular',
} as const;

export type ListReviewsParams = {
productId?: number;
};