# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Flower delivery shop API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Flower product catalog
  - name: cart
    description: Shopping cart
  - name: orders
    description: Order management
  - name: ai
    description: AI bouquet assistant
  - name: reviews
    description: Product reviews
  - name: catalog
    description: Catalog aggregates

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, name, popular]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/bestseller products for homepage
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /catalog/summary:
    get:
      operationId: getCatalogSummary
      tags: [catalog]
      summary: Get catalog summary - categories, price range, total count
      responses:
        "200":
          description: Catalog summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogSummary"

  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order status and details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews (optionally by product)
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /ai/recommend:
    post:
      operationId: aiRecommend
      tags: [ai]
      summary: Get AI bouquet recommendation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiRecommendRequest"
      responses:
        "200":
          description: AI recommendation response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiReply"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        inStock:
          type: boolean
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
      required:
        - id
        - name
        - description
        - price
        - imageUrl
        - category
        - inStock
        - isFeatured
        - rating
        - reviewCount

    ProductInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        category:
          type: string
        inStock:
          type: boolean
        isFeatured:
          type: boolean
      required:
        - name
        - description
        - price
        - imageUrl
        - category

    CatalogSummary:
      type: object
      properties:
        totalProducts:
          type: integer
        categories:
          type: array
          items:
            type: string
        minPrice:
          type: number
        maxPrice:
          type: number
        featuredCount:
          type: integer
      required:
        - totalProducts
        - categories
        - minPrice
        - maxPrice
        - featuredCount

    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - productId
        - productName
        - quantity
        - price

    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
        deliveryTime:
          type: string
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, delivering, delivered, cancelled]
        totalPrice:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - customerEmail
        - deliveryAddress
        - deliveryDate
        - deliveryTime
        - status
        - totalPrice
        - items
        - createdAt

    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity

    OrderInput:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: string
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
        deliveryTime:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required:
        - customerName
        - customerPhone
        - customerEmail
        - deliveryAddress
        - deliveryDate
        - deliveryTime
        - items

    Review:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerName
        - rating
        - comment
        - createdAt

    ReviewInput:
      type: object
      properties:
        productId:
          type: ["integer", "null"]
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - customerName
        - rating
        - comment

    AiRecommendRequest:
      type: object
      properties:
        message:
          type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - message

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    AiReply:
      type: object
      properties:
        reply:
          type: string
        recommendedProductIds:
          type: array
          items:
            type: integer
      required:
        - reply
        - recommendedProductIds