# lib/api-zod/src/generated/types/aiRecommendRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flower delivery shop API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from './chatMessage';

export interface AiRecommendRequest {
  message: string;
  history?: ChatMessage[];
}