# lib/api-zod/src/generated/types/catalogSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flower delivery shop API
 * OpenAPI spec version: 0.1.0
 */

export interface CatalogSummary {
  totalProducts: number;
  categories: string[];
  minPrice: number;
  maxPrice: number;
  featuredCount: number;
}