# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flower delivery shop API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItemInput } from './orderItemInput';

export interface OrderInput {
  customerName: string;
  customerPhone: string;
  customerEmail: string;
  deliveryAddress: string;
  deliveryDate: string;
  deliveryTime: string;
  notes?: string;
  items: OrderItemInput[];
}