# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Flower delivery shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: number;
  /** @nullable */
  productId?: number | null;
  customerName: string;
  rating: number;
  comment: string;
  createdAt: string;
}